Eliminate AbstractTestRunner class
